Add Birb class
Class for entity added, followed by reformat of the code
